+++ tests/jest-fastify-demo/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @nodeboot/jest-fastify-demo
 
+## 1.2.1
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/fastify-server@1.7.0
+    -   @nodeboot/starter-actuator@1.6.0
+    -   @nodeboot/starter-openapi@2.3.0
+
 ## 1.2.0
 
 ### Minor Changes